api/requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
psycopg2==2.8.5
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
